+++ src/dynamic.ts
@@ -53,7 +53,6 @@ export function addPlatformSpecificChanges(uePlatform: UnrealPlatform, clangdExt
             break;
         case 'Linux':
             addToClangdAdd(clangdExtYamlFiles, consts.LINUX_CLANGD_CFG_ADD_USR_INCLUDE);
-			addToClangdAdd(clangdExtYamlFiles, consts.LINUX_CLANGD_CFG_ADD_PROMOTE_PREINCLUDE);
             break;
         default:
             break;